Add option for dynamic attrs with suggested changes
