+++ src/mocker.js
@@ -1,6 +1,6 @@
 const { loadAndSetUserConfigurations, config } = require('./config');
 const { processVocabulary } = require('./vocabulary');
-const { generateCases, generateEvents, generateDynamicCaseAttrs } = require('./data');
+const { generateCases, generateEvents } = require('./data');
 const { saveToCSV, writeFile } = require('./output');
 const { generateSchemaSql, generateSqlInsert } = require('./sql_generator');
 const util = require('./util');
@@ -8,9 +8,7 @@ const pluralize = require('pluralize');
 
 async function main() {
   loadAndSetUserConfigurations();
-
-  const dynamicCaseAttrs = generateDynamicCaseAttrs(config.DYNAMIC_ATTRS, config.UNIQUE_VALUES_FOR_DYNAMIC_ATTRS);
-  const vocabulary = processVocabulary(dynamicCaseAttrs);
+  const vocabulary = processVocabulary();
 
   const cases = generateCases(config.NUMBER_OF_CASES);
   const events = generateEvents(cases);
@@ -36,7 +34,6 @@ async function main() {
 
 function saveToSql(vocabulary, cases, events) {
   let combinedFile = '';
-  let additionalCaseAttrs = 10
 
   const schema = generateSchemaSql(vocabulary.schema);
   writeFile('out/schema.sql', schema);
